@@ -2,6 +2,7 @@
   "name": "MillyardPong",
   "description": "Millyard Pong game",
   "version": "1.0.0",
+  "type": "module",
   "author": "Zack Mattor <zmattor@gmail.com>",
   "private": true,
   "scripts": {
@@ -10,27 +11,28 @@
     "build": "./bin/build.sh"
   },
   "dependencies": {
-    "express": "^4.16.4",
-    "jquery": "^3.2.1",
-    "vue": "^2.6.6",
-    "vue-analytics": "^4.1.2",
-    "vue-router": "^2.5.3",
-    "ws": "^1.1.1"
+    "@laser-dac/core": "^0.4.2",
+    "@laser-dac/laserdock": "^0.4.2",
+    "@laser-dac/simulator": "^0.3.4",
+    "babel-core": "^6.26.3",
+    "babel-preset-env": "^1.7.0",
+    "express": "^4.18.1",
+    "isomorphic-ws": "^5.0.0",
+    "vue": "^3.2.39",
+    "vue-router": "^4.1.5",
+    "vue-style-loader": "^4.1.3"
   },
   "devDependencies": {
-    "babel-core": "^6.26.3",
-    "babel-loader": "^6.0.0",
-    "babel-preset-latest": "^6.0.0",
-    "copy-webpack-plugin": "^4.6.0",
-    "cross-env": "^3.0.0",
-    "css-loader": "^0.25.0",
-    "file-loader": "^0.9.0",
-    "node-sass": "^4.11.0",
-    "sass-loader": "^5.0.1",
-    "style-loader": "^0.18.1",
-    "vue-loader": "^11.1.4",
-    "vue-template-compiler": "^2.6.6",
-    "webpack": "^2.2.0",
-    "webpack-dev-server": "^2.11.3"
+    "babel-loader": "^8.2.5",
+    "copy-webpack-plugin": "^11.0.0",
+    "css-loader": "^6.7.1",
+    "jquery": "^3.6.1",
+    "sass": "^1.55.0",
+    "sass-loader": "^13.0.2",
+    "style-loader": "^3.3.1",
+    "vue-loader": "^17.0.0",
+    "webpack": "^5.74.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-server": "^4.11.1"
   }
 }
