@@ -1,11 +1,11 @@
-require('../shared/style.css');
+import '../shared/style.css';
 
-import Vue from 'vue';
-import VueRouter from 'vue-router';
-import GameConnection from '../shared/game_connection.js';
+import { createApp } from 'vue';
+import { createRouter, createWebHashHistory } from 'vue-router';
+import { GameConnection } from '../shared/game_connection.js';
+
 import jquery from 'jquery';
-import VueAnalytics from 'vue-analytics';
-
+import IndexComponent from './root.vue';
 import IndexRoute from './routes/index/index.vue';
 import GameOverRoute from './routes/game-over/index.vue';
 import GameRoute from './routes/game/index.vue';
@@ -21,41 +21,21 @@ const routes = [
   { path: '/game-over', component: GameOverRoute }
 ];
 
-const router = new VueRouter({
+const router = createRouter({
+  history: createWebHashHistory(),
   routes
 });
 
-
-Vue.use(VueRouter);
-
-if(window.location.origin === "http://mhtpong.com") {
-  Vue.use(VueAnalytics, {
-    id: 'UA-100333942-1',
-    router
-  });
-} else {
-  // Mock vue analytics
-  Vue.use({
-    install(Vue, opts) {
-      Vue.prototype.$ga = Vue.$ga = {
-        event(name, val) {
-          console.log(` Mock GA Event -> ${name} - ${val}`)
-        }
-      }
-    }
-  });
-}
 
 var Controller = {
   init() {
     this.disable_zoom();
 
-    Vue.prototype.$game_connection = new GameConnection('ws_controller');
-    Vue.prototype.$store = {};
-
-    let index_component = require('./root.vue');
-
-    const app = new Vue(Object.assign(index_component, {router: router})).$mount('#app');
+    const app = createApp(IndexComponent);
+    app.provide('$game_connection', new GameConnection('ws_controller'));
+    app.provide('$store', {});
+    app.use(router);
+    app.mount("#app");
 
     router.replace('/')
 
