@@ -1,9 +1,9 @@
-const WebSocket = require('ws');
-const http = require('http');
-const ServerManager = require('./server/server_manager.js');
+import { WebSocketServer } from 'ws';
+import http from 'http';
+import { ServerManager } from './server/server_manager.js';
 
 const server = http.createServer();
-const wss = new WebSocket.Server({ server });
+const wss = new WebSocketServer({ server });
 
 let port = process.env.MANCHESTER_PONG_PORT || 3000;
 
@@ -11,4 +11,4 @@ server.listen(port, () => {
   console.log(`Example app listening on port ${port}!`);
 });
 
-let pong_server = new ServerManager(wss);
+(new ServerManager(wss))
