@@ -11,13 +11,14 @@
          v-on:touchstart="pressed('right')"
          v-on:touchend="released('right')"
          class="right"><i class="fa fa-arrow-right" aria-hidden="true"></i></div>
-    
   </div>
 </template>
 
 <script>
 export default {
+  inject: ['$game_connection', '$store'],
   name: 'game',
+
   mounted() {
     this.$game_connection.on_game_over = (data) => {
       if(data.err) {
