@@ -15,6 +15,7 @@
 import Keypad from './keypad.vue';
 
 export default {
+  inject: ['$game_connection', '$store'],
   name: 'index',
 
   mounted() {
@@ -41,14 +42,12 @@ export default {
     on_bind_status(data) {
       if(!data.was_successful) {
         this.err = 'Invalid Token...';
-        this.$ga.event('bind_status', 'failed');
 
         setTimeout(() => {
           this.err = '';
           this.token = '';
         }, 1300);
       } else {
-        this.$ga.event('bind_status', 'success');
         this.$store.player_id = data.player_id;
         this.$router.push('/lobby');
       }
