@@ -1,15 +1,16 @@
 var path = require('path')
 var webpack = require('webpack')
 var CopyWebpackPlugin = require('copy-webpack-plugin');
+const { VueLoaderPlugin } = require('vue-loader')
 
 module.exports = {
+  mode: 'development',
   entry: {
     controller: './src/mobile/index.js',
     game: './src/game/index.js'
   },
   output: {
     path: path.resolve(__dirname, './dist/assets'),
-    //publicPath: '/dist/assets',
     filename: '[name].js'
   },
   module: {
@@ -17,17 +18,10 @@ module.exports = {
       {
         test: /\.vue$/,
         loader: 'vue-loader',
-        options: {
-          loaders: {
-            'scss': 'vue-style-loader!css-loader!sass-loader',
-            'sass': 'vue-style-loader!css-loader!sass-loader?indentedSyntax'
-          }
-        }
       },
       {
         test: /\.js$/,
         loader: 'babel-loader',
-        exclude: /node_modules/
       },
       {
         test: /\.(png|jpg|gif|svg)$/,
@@ -37,34 +31,39 @@ module.exports = {
         }
       },
       {
-        test: /\.css$/,
-        use: [ 'style-loader', 'css-loader' ]
-      }
+        test: /\.(scss|css)$/,
+        use: [
+          'style-loader',
+          {loader: 'css-loader', options: {sourceMap: true, importLoaders: 1}},
+          {loader: 'sass-loader', options: {sourceMap: true}},
+        ],
+      },
     ]
   },
   plugins: [
-    new CopyWebpackPlugin([
-      { from: 'src/mobile.html', to: '../index.html' },
-      { from: 'src/web_display.html', to: '../game/index.html' },
-      { from: 'src/web_display_test.html', to: '../test.html' },
-    ])
+    new VueLoaderPlugin(),
+    new CopyWebpackPlugin(
+      {
+        patterns: [
+          { from: 'src/mobile.html', to: '../index.html' },
+          { from: 'src/web_display.html', to: '../game/index.html' },
+          { from: 'src/web_display_test.html', to: '../test.html' },
+        ]
+      }
+    )
   ],
-  resolve: {
-    alias: {
-      'vue$': 'vue/dist/vue.esm.js'
-    }
-  },
   devServer: {
     proxy : [{path: '/ws_game', target: 'ws://localhost:3000', ws: true}, {path: '/ws_controller', target: 'ws://localhost:3000', ws: true}],
     historyApiFallback: true,
-    noInfo: true,
-    publicPath: 'dist',
-    contentBase: 'dist'
+    devMiddleware: {
+      publicPath: 'dist'
+    },
+    static: 'dist'
   },
   performance: {
     hints: false
   },
-  devtool: '#eval-source-map'
+  devtool: 'source-map'
 }
 
 if (process.env.NODE_ENV === 'production') {
